# contracts/v1/openapi.yaml
openapi: "3.1.0"
info:
  title: Cruzonic Fleet Management API
  version: "1.0.0"
  description: |
    Versioned REST API contract for the Cruzonic Fleet Management Platform.

    **Locked version:** v1.0.0
    **Contract registry:** /contracts/README.md

    ## Authentication
    All endpoints require a Supabase-issued JWT passed as a Bearer token.
    The JWT encodes `sub` (user UUID), `role`, and `carrier_id` claims.
    The backend validates the token using the Supabase JWT secret.

    ## Versioning
    Include `X-Contract-Version: 1.0.0` on every request.
    The server logs a warning (but does not reject) requests missing this header.

    ## Rate limits
    - Event ingestion: 500 events/minute per device
    - Query endpoints: 120 requests/minute per user
    - Compliance export: 10 requests/minute per carrier
    Rate limit state is returned in `X-RateLimit-*` response headers.

    ## Error format
    All errors return `application/json` with the `ApiError` schema (see components).

  contact:
    name: Cruzonic Engineering
    email: engineering@cruzonic.io
  license:
    name: Proprietary
servers:
  - url: https://api.cruzonic.io/v1
    description: Production
  - url: https://api.staging.cruzonic.io/v1
    description: Staging
  - url: http://localhost:3000/v1
    description: Local development

# ─────────────────────────────────────────────────────────────────────────────
# SECURITY
# ─────────────────────────────────────────────────────────────────────────────
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase-issued JWT. Obtain via POST /auth/v1/token (Supabase Auth API).
  headers:
...
    X-RateLimit-Limit:
      description: Maximum number of requests allowed in the current window.
      schema:
        type: integer
    X-RateLimit-Remaining:
      description: Remaining requests in the current window.
      schema:
        type: integer
...
  parameters:
    PageParam:
      name: page
      in: query
      description: 1-based page number for paginated responses.
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Maximum number of items per page.
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code.
          examples: [VALIDATION_ERROR, RATE_LIMITED, UNAUTHORIZED, NOT_FOUND, SEQUENCE_DUPLICATE]
        message:
          type: string
          description: Human-readable explanation.
        details:
          type: array
          description: Field-level validation errors (present for VALIDATION_ERROR only).
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
                description: Dot-notation path to the invalid field.
              message:
                type: string
        requestId:
          type: string
          description: Unique ID for this request — include in bug reports.

    # ── Pagination envelope ───────────────────────────────────────────────────
    PaginationMeta:
      type: object
      required: [page, limit, total, totalPages]
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
          description: Total number of matching records across all pages.
        totalPages:
          type: integer

    # ── ELD event (shared fields) ─────────────────────────────────────────────
    EldEventCommon:
      type: object
      description: Fields present on every ELD event regardless of type.
      required:
        - eventSequenceId
        - eventRecordStatus
        - eventRecordOrigin
        - eventType
        - eventSubType
        - eventDate
        - eventTime
        - timezoneOffset
        - accumulatedVehicleMiles
        - elapsedEngineHours
        - location
        - vehicle
        - driver
        - eldRegistrationId
        - malfunctionIndicatorStatus
        - eldMalfunctionStatus
        - dataDiagnosticIndicatorStatus
        - eventDataCheckValue
      properties:
        eventSequenceId:
          type: integer
          minimum: 1
          maximum: 65535
          description: Monotonically increasing per (eldDeviceId × logDate). Resets at midnight home-terminal time.
        eventRecordStatus:
          type: integer
          enum: [1, 2, 3, 4]
          description: "1=Active 2=Inactive-Changed 3=Inactive-ChangeRequested 4=Inactive-AssumedUnidentified"
        eventRecordOrigin:
          type: integer
          enum: [1, 2, 3, 4]
          description: "1=AutoRecorded 2=DriverEdit 3=OtherUserEdit 4=UnidentifiedDriver"
        eventType:
          type: integer
          enum: [1, 2, 3, 4, 5, 6, 7]
          description: "1=DutyStatus 2=IntermediateLog 3=PersonalUse/YardMoves 4=Certification 5=Login/Logout 6=EnginePower 7=Malfunction/Diagnostic"
        eventSubType:
          type: integer
          minimum: 1
          description: Secondary classification — interpretation depends on eventType.
        eventDate:
          type: string
          pattern: '^(0[1-9]|1[0-2])(0[1-9]|[12][0-9]|3[01])[0-9]{2}$'
          description: Calendar date MMDDYY in home-terminal timezone.
          examples: ["061224"]
        eventTime:
          type: string
          pattern: '^([01][0-9]|2[0-3])[0-5][0-9][0-5][0-9]$'
          description: Event time HHMMSS in UTC.
          examples: ["140000"]
        timezoneOffset:
          type: string
          pattern: '^[+-](0[0-9]|1[0-4])[0-5][0-9]$'
          description: Home-terminal UTC offset (±HHMM).
          examples: ["-0600"]
        accumulatedVehicleMiles:
          type: number
          minimum: 0
          description: Engine-reported odometer in miles (1 decimal).
        elapsedEngineHours:
          type: number
          minimum: 0
          description: Engine-reported total hours (1 decimal).
        location:
          $ref: '#/components/schemas/LocationPayload'
        vehicle:
          $ref: '#/components/schemas/VehiclePayload'
        driver:
          $ref: '#/components/schemas/DriverPayload'
        eldRegistrationId:
          type: string
          maxLength: 8
          description: FMCSA-assigned ELD device registration ID.
          examples: ["PACTRKV1"]
        malfunctionIndicatorStatus:
          type: boolean
        eldMalfunctionStatus:
          type: boolean
        dataDiagnosticIndicatorStatus:
          type: boolean
        eventDataCheckValue:
          type: string
          pattern: '^[0-9A-Fa-f]{2}$'
          description: Two-hex-char FMCSA checksum (Appendix A §7.3.2).
        annotation:
          type: string
          maxLength: 60
          description: Optional driver annotation.
    LocationPayload:
      type: object
      required: [latitude, longitude, locationDescription, distanceSinceLastValidCoordinates]
      properties:
        latitude:
          type: number
          minimum: -90
          maximum: 90
          nullable: true
          description: WGS-84 decimal degrees. Null when GPS unavailable.
        longitude:
          type: number
          minimum: -180
          maximum: 180
          nullable: true
        locationDescription:
          type: string
          maxLength: 60
          description: Nearest city/state description.
          examples: ["Chicago, IL"]
        distanceSinceLastValidCoordinates:
          type: number
          minimum: 0
          nullable: true
          description: Tenths of a mile since last GPS fix. Required when lat/lon is null.
    VehiclePayload:
      type: object
      required: [cmvPowerUnitNumber, vin]
      properties:
        cmvPowerUnitNumber:
          type: string
          maxLength: 10
        vin:
          type: string
          minLength: 17
          maxLength: 17
          pattern: '^[A-HJ-NPR-Z0-9]{17}$'
        licensePlate:
          type: string
          maxLength: 12
        licensePlateState:
          type: string
          pattern: '^[A-Z]{2}$'
    DriverPayload:
      type: object
      required: [driverEldAccountId, carrierDotNumber]
      properties:
        driverEldAccountId:
          type: string
          maxLength: 60
        coDriverEldAccountId:
          type: string
          maxLength: 60
        carrierDotNumber:
          type: string
          pattern: '^[0-9]{1,8}$'
    EldEventResponse:
      allOf:
        - $ref: '#/components/schemas/EldEventCommon'
      type: object
      required: [id, logPeriodId, carrierId, createdAt, versionNumber, originalVersionId]
      properties:
        id:
          type: string
          format: uuid
        logPeriodId:
          type: string
          format: uuid
        carrierId:
          type: string
          format: uuid
        versionNumber:
          type: integer
          minimum: 1
        previousVersionId:
          type: string
          format: uuid
          nullable: true
        originalVersionId:
          type: string
          format: uuid
        contentHash:
          type: string
          description: SHA-256 hex of canonical event data fields.
        chainHash:
          type: string
          description: SHA-256 chain hash linking this record to its predecessor.
        createdAt:
          type: string
          format: date-time

    # ── Log period ─────────────────────────────────────────────────────────────
...
    SyncBatchRequest:
      type: object
      required: [deviceId, events, syncedUpToAt]
      properties:
        deviceId:
          type: string
          description: ELD device registration ID or mobile device identifier.
        syncedUpToAt:
          type: string
          format: date-time
          description: The client's last successful sync timestamp. The server returns any server-side changes since this point.
        events:
          type: array
          maxItems: 500
          description: Ordered array of buffered ELD events (earliest first).
          items:
            $ref: '#/components/schemas/EldEventCommon'
    SyncBatchResponse:
      type: object
      required: [accepted, rejected, serverEvents, newSyncedUpToAt]
      properties:
        accepted:
          type: array
          description: IDs of events that were successfully ingested.
          items:
            type: integer
            description: eventSequenceId of the accepted event.
        rejected:
          type: array
          description: Events that were rejected, with reasons.
          items:
            type: object
            required: [eventSequenceId, errors]
            properties:
              eventSequenceId:
                type: integer
              errors:
                type: array
                items:
                  type: object
                  required: [code, message]
                  properties:
                    code:
                      type: string
                      enum:
                        - DUPLICATE
                        - NON_MONOTONIC
                        - OUT_OF_RANGE
                        - INVALID_CHECKSUM
                        - SCHEMA_VALIDATION_ERROR
                        - WRONG_SCOPE
                    message:
                      type: string
        serverEvents:
          type: array
          description: ELD events that originated on the server side (carrier edits, etc.) since syncedUpToAt. The client must merge these.
          items:
            $ref: '#/components/schemas/EldEventResponse'
        newSyncedUpToAt:
          type: string
          format: date-time
          description: Client must persist this and send it in the next sync request.

# ─────────────────────────────────────────────────────────────────────────────
# PATHS
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /eld/events:
    post:
      operationId: ingestEldEvents
      summary: Ingest a batch of ELD events
      description: |
        Accepts up to 100 ELD events per request. Events must be ordered by
        `eventSequenceId` ascending within each (eldDeviceId × logDate) scope.

        **Partial acceptance:** The server processes each event independently.
        Events that fail validation are rejected with per-event error details;
        valid events in the same batch are still accepted. The response lists
        all accepted and rejected sequence IDs.

        **Idempotency:** Duplicate events (same eldDeviceId + logDate + eventSequenceId
        with identical content) are silently accepted and counted as successful
        (the server returns the existing record ID). Duplicates with *different*
        content are rejected with code `DUPLICATE`.

        **Rate limit:** 500 events per minute per device.
      tags: [ELD Events]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [events]
              properties:
                events:
                  type: array
                  minItems: 1
                  maxItems: 100
                  items:
                    $ref: '#/components/schemas/EldEventCommon'
            examples:
              dutyStatusChange:
                summary: Single duty-status change (driver started driving)
                value:
                  events:
                    - eventSequenceId: 5
                      eventRecordStatus: 1
                      eventRecordOrigin: 1
                      eventType: 1
                      eventSubType: 3
                      eventDate: "061224"
                      eventTime: "140000"
                      timezoneOffset: "-0600"
                      accumulatedVehicleMiles: 48321.4
                      elapsedEngineHours: 4521.2
                      location:
                        latitude: 41.878114
                        longitude: -87.629799
                        locationDescription: "Chicago, IL"
                        distanceSinceLastValidCoordinates: null
                      vehicle:
                        cmvPowerUnitNumber: "TRUCK-042"
                        vin: "1HGBH41JXMN109186"
                      driver:
                        driverEldAccountId: "jsmith"
                        carrierDotNumber: "1234567"
                      eldRegistrationId: "PACTRKV1"
                      malfunctionIndicatorStatus: false
                      eldMalfunctionStatus: false
                      dataDiagnosticIndicatorStatus: false
                      eventDataCheckValue: "3C"
                      driverEldAccountId: "jsmith"
      responses:
        "207":
          description: Multi-status — some events accepted, some rejected.
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
          content:
            application/json:
              schema:
                type: object
                required: [accepted, rejected]
                properties:
                  accepted:
                    type: array
                    items:
                      type: object
                      required: [eventSequenceId, id]
                      properties:
                        eventSequenceId:
                          type: integer
                        id:
                          type: string
                          format: uuid
                          description: Server-assigned UUID for this event record.
                  rejected:
                    type: array
                    items:
                      type: object
                      required: [eventSequenceId, errors]
                      properties:
                        eventSequenceId:
                          type: integer
                        errors:
                          type: array
                          items:
                            $ref: '#/components/schemas/ApiError'
        "401":
          description: Missing or invalid JWT.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        "429":
          description: Rate limit exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

    get:
      operationId: queryEldEvents
      summary: Query ELD events
      description: |
        Returns ELD event records for the authenticated user's scope.
        Drivers see only their own events; fleet managers see their carrier's events.

        **Default:** returns only active records (`eventRecordStatus=1`).
        Pass `includeInactive=true` to include superseded versions.
      tags: [ELD Events]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: driverId
          in: query
          schema:
            type: string
            format: uuid
        - name: logPeriodId
          in: query
          schema:
            type: string
            format: uuid
        - name: eventType
          in: query
          description: Filter by event type (1–7). Repeatable for multiple types.
          schema:
            type: array
            items:
              type: integer
              minimum: 1
              maximum: 7
          style: form
          explode: true
        - name: eventDate
          in: query
          description: Filter by log date (MMDDYY).
          schema:
            type: string
            pattern: '^(0[1-9]|1[0-2])(0[1-9]|[12][0-9]|3[01])[0-9]{2}$'
        - name: fromDate
          in: query
          description: Filter events on or after this ISO 8601 date (inclusive).
          schema:
            type: string
            format: date
        - name: toDate
          in: query
          description: Filter events on or before this ISO 8601 date (inclusive).
          schema:
            type: string
            format: date
        - name: includeInactive
          in: query
          description: When true, includes superseded (inactive) versions of edited records.
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Paginated list of ELD events.
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EldEventResponse'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /eld/events/{eventId}:
    get:
      operationId: getEldEvent
      summary: Get a single ELD event by ID
      description: Returns the active version of an event. Includes its full edit history if `includeHistory=true`.
      tags: [ELD Events]
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: includeHistory
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EldEventResponse'
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  # ── Sequence ID ─────────────────────────────────────────────────────────────
...
  /sync/events:
    post:
      operationId: syncEvents
      summary: Offline event sync (drain offline buffer)
      description: |
        Drains the mobile app's offline event buffer to the server in a single
        atomic request. See `/contracts/v1/sync-protocol.md` for the full
        protocol specification including conflict resolution and retry strategy.

        Events must be sorted by `(logDate ASC, eventSequenceId ASC)`.
        Maximum 500 events per request.
      tags: [Sync]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncBatchRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncBatchResponse'
        "400":
          description: Batch validation error (e.g. events not ordered correctly).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...